robotic_arm/pincherx100/camera_calibration/requirements.txt
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
flask>=3.0.0
pillow>=10.0.0

